+++ app/src/main/java/com/ze20/saifu/ui/log/RecyclerView/LogRowModel.kt
@@ -1,8 +1,8 @@
-package com.ze20.saifu.ui.setting.Recyclerview
+package com.ze20.saifu.ui.log.RecyclerView
 
 class LogRowModel {
     // リストのアイテムを表すクラス
-    var name: String = ""
+    var day: String = ""
     var category: String = ""
     var price: String = ""
 }
