+++ app/src/main/java/com/ze20/saifu/ui/log/RecyclerView/RecyclerViewHolder.kt
@@ -1,4 +1,4 @@
-package com.ze20.saifu.ui.log.Recyclerview
+package com.ze20.saifu.ui.log.RecyclerView
 
 import android.view.View
 import android.widget.TextView
