+++ app/src/main/java/com/ze20/saifu/ui/log/RecyclerView/ViewAdapter.kt
@@ -1,4 +1,4 @@
-package com.ze20.saifu.ui.setting.Recyclerview
+package com.ze20.saifu.ui.log.RecyclerView
 
 import android.util.Log
 import android.view.LayoutInflater
@@ -6,14 +6,13 @@ import android.view.View
 import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
 import com.ze20.saifu.R
-import com.ze20.saifu.ui.Recyclerview.ShortcutRowModel
 
 // ViewHolderを生成しViewModelをセットする
 // 詳しくはココ参照→ https://qiita.com/saiki-ii/items/78ed73134784f3e5db7e
 
-class ShortcutViewAdapter(
+class ViewAdapter(
 
-    private val list: List<ShortcutRowModel>,
+    private val list: List<LogRowModel>,
     private val listener: ListListener
 ) : RecyclerView.Adapter<RecyclerViewHolder>() {
 
@@ -21,21 +20,15 @@ class ShortcutViewAdapter(
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): RecyclerViewHolder {
         Log.d("Adapter", "onCreateViewHolder")
         val rowView: View =
-            LayoutInflater.from(parent.context)
-                .inflate(R.layout.activity_edit_shortcut_list, parent, false)
+            LayoutInflater.from(parent.context).inflate(R.layout.fragment_log_list, parent, false)
         return RecyclerViewHolder(rowView)
     }
 
     // positionをlistのindexとしてRecyclerViewHolderに値をセットする
     override fun onBindViewHolder(holder: RecyclerViewHolder, position: Int) {
-        if (list[position].name == "" || list[position].name == " " || list[position].name.isEmpty()) {
-            holder.mainView.text = list[position].price
-            holder.subView.visibility = View.GONE
-        } else {
-            holder.mainView.text = list[position].name
-            holder.subView.text = list[position].price
-        }
-        // holder.categoryView.text = list[position].category
+        holder.dayView.text = list[position].day
+        holder.categoryView.text = list[position].category
+        holder.priceView.text = list[position].price
         holder.itemView.setOnClickListener {
             listener.onClickRow(it, list[position])
         }
@@ -47,6 +40,6 @@ class ShortcutViewAdapter(
 
     // インターフェイースの生成
     interface ListListener {
-        fun onClickRow(tappedView: View, rowModel: ShortcutRowModel)
+        fun onClickRow(tappedView: View, rowModel: LogRowModel)
     }
 }
